refactored code base to use extension in own file
